% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_fahrenheit.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Convert temperature from Celsius to Fahrenheit}
\usage{
celsius_to_fahrenheit(celsius)
}
\arguments{
\item{celsius}{A number}
}
\value{
Number representing temperature in fahrenheit
}
\description{
Given an input temperature number, this function would convert
it from celsius to fahrenheit.
}
\examples{
celsius_to_fahrenheit(10)
celsius_to_fahrenheit(0)
}
