% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_kelvin.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{Convert temperature from Celsius to Kelvin}
\usage{
celsius_to_kelvin(celsius)
}
\arguments{
\item{celsius}{A number}
}
\value{
Number representing temperature in kelvin
}
\description{
Given an input temperature number, this function would convert
it from celsius to kelvin.
}
\examples{
celsius_to_kelvin(10)
celsius_to_kelvin(0)
}
