% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert temperature from Fahrenheit to celsius}
\usage{
fahrenheit_to_celsius(fahrenheit)
}
\arguments{
\item{fahrenheit}{A number}
}
\value{
Number representing temperature in celsius
}
\description{
Given an input temperature number, this function would convert
it from fahrenheit to celsius.
}
\examples{
fahrenheit_to_celsius(10)
fahrenheit_to_celsius(0)
}
