% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_kelvin.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Convert temperature from Fahrenheit to kelvin}
\usage{
fahrenheit_to_kelvin(fahrenheit)
}
\arguments{
\item{fahrenheit}{A number}
}
\value{
Number representing temperature in kelvin
}
\description{
Given an input temperature number, this function would convert
it from fahrenheit to kelvin.
}
\examples{
fahrenheit_to_kelvin(10)
fahrenheit_to_kelvin(0)
}
