% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kevin_to_celsius.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{Convert temperature from kelvin to celsius}
\usage{
kelvin_to_celsius(kelvin)
}
\arguments{
\item{kelvin}{A number}
}
\value{
Number representing temperature in celsius
}
\description{
Given an input temperature number, this function would convert
it from kelvin to celsius.
}
\examples{
kelvin_to_celsius(10)
kelvin_to_celsius(0)
}
