% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kevin_to_fahrenheit.R
\name{kelvin_to_fahrenheit}
\alias{kelvin_to_fahrenheit}
\title{Convert temperature from kelvin to fahrenheit}
\usage{
kelvin_to_fahrenheit(kelvin)
}
\arguments{
\item{kelvin}{A number}
}
\value{
Number representing temperature in fahrenheit
}
\description{
Given an input temperature number, this function would convert
it from kelvin to fahrenheit.
}
\examples{
kelvin_to_fahrenheit(10)
kelvin_to_fahrenheit(0)
}
